feat(apama_project): create project in current dir (#67)
we only want a single project per workspace directory, so we now only support creating a project within the current workspace directory.

this feature relies on an apama change.

we also fix the messaging, so that success & error command lines are added as information messages (in addition to being logged).