fix(apamaProjectView): remove deployment support (#72)
We currently don't support deployment in a sub-directory of the current
project directory (ironically, this was something we deliberately axed
due to bugs in that mechanism). This makes it awkward to fit deployment
into the developer story when using VS Code, because ideally, the build
output would be a sub-directory of the current working directory.

For now, we're removing the support for deployment. Alternatives for the
future could include,

- Adding a dialog to allow the user to select the deployment directory, 
- Deploying it to a default directory outside of the workspace (i.e.
`../<PROJECT_NAME>_deployed`),
- Implementing support for deployment into a subdirectory,

(Ideally, we'd go with 3), and add some mechanism to avoid recursive
deployments like last time).