test: initial attempt to expand testing
in this commit, we attempt to add testing for validating that the
commands are successfully added to the vscode command palette.

this works when invoked inside of a properly configured environment,
however, when ran outside of it, it fails because the code we're testing
is executed after we verify if we know where Apama is.

there isn't really much to test before Apama is detected, so this is the
critical blocker for progressing further.